refactor: Move null comparison validation to SQLValidator in analysis phase
- Created dedicated SQLValidator class in the analyzer package
- Moved null comparison warning logic from RewriteExpr to SQLValidator
- Added SQLValidator to analysis phases after TypeResolver
- Removed WarnNullComparison from expression rewrite rules

This architectural change improves separation of concerns:
- Validation logic is now in the analysis phase where it belongs
- SQL-specific checks are isolated in a dedicated validator
- Transform phase focuses only on expression transformations
