+++ wvlet-lang/src/main/scala/wvlet/lang/compiler/Compiler.scala
@@ -20,6 +20,7 @@ import wvlet.lang.compiler.analyzer.{
   EmptyTypeResolver,
   ModelDependencyAnalyzer,
   RemoveUnusedQueries,
+  SQLValidator,
   SymbolLabeler,
   TypeResolver
 }
@@ -54,7 +55,8 @@ object Compiler extends LogSupport:
     PreprocessLocalExpr, // Preprocess local expressions (e.g., backquote strings and native expressions)
     SymbolLabeler, // Assign unique Symbol to each LogicalPlan and Expression nodes, a and assign a lazy DataType
     RemoveUnusedQueries(), // Exclude unused compilation units (e.g., out of scope queries) from the following phases
-    TypeResolver, // Assign a concrete DataType to each LogicalPlan and Expression nodes
+    TypeResolver,   // Assign a concrete DataType to each LogicalPlan and Expression nodes
+    SQLValidator(), // Validate SQL-specific patterns and emit warnings
     ModelDependencyAnalyzer()
   )
